+++ trustpinning/certs_test.go
@@ -1,4 +1,4 @@
-package certs
+package trustpinning
 
 import (
 	"bytes"
