+++ trustpinning/trustpin.go
@@ -1,4 +1,4 @@
-package certs
+package trustpinning
 
 import (
 	"crypto/x509"
@@ -42,14 +42,13 @@ func NewTrustPinChecker(trustPinConfig notary.TrustPinConfig, gun string) (Trust
 		if err != nil {
 			return TrustPinChecker{}, fmt.Errorf("could not load root cert from CA path")
 		}
-		if err = trustmanager.ValidateCertificate(caCerts); err != nil {
-			return TrustPinChecker{}, fmt.Errorf("invalid CA cert provided")
-		}
 		// Now only consider certificates that are direct children from this CA cert, overwriting allValidCerts
 		caRootPool := x509.NewCertPool()
-		caRootPool.AddCert(caCerts)
-		if err != nil {
-			return TrustPinChecker{}, fmt.Errorf("unable to initialize CA cert pool")
+		for _, caCert := range caCerts {
+			if err = trustmanager.ValidateCertificate(caCert); err != nil {
+				return TrustPinChecker{}, fmt.Errorf("invalid CA cert provided")
+			}
+			caRootPool.AddCert(caCert)
 		}
 		trustPinChecker.pinnedCAPool = caRootPool
 		return trustPinChecker, nil
